Netty fixed readTimeout
Added local ReadTimeoutHandler based off the netty handler that
measures the idle time between read start and read completed. This is
inline with the AsynchronousSocketChannelStream that has a per read
request timeout.

The Netty ReadTimeoutHandler measured idle time between any reads
happening in the channel. This handler would throw an exception and close
the channel if there was a pause in asking for a read.

JAVA-1934