Docs API header updates
Ensures the location hash is passed to the "classFrames" frame. This is
done in a try / catch due so doesn't error due to same origin policy.

Added google analytics tag.  Have to check if we are in frames view, then
the only page we care about is the "classFrame" or in no frames view then
we always care.

JAVA-1724